% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_directory.R
\name{abs_directory}
\alias{abs_directory}
\title{Simple wrapper to abs_path to accomodate symlinks}
\usage{
abs_directory(x)
}
\arguments{
\item{x}{`character` path to file.}
}
\description{
Simple wrapper to abs_path to accomodate symlinks
}
\details{
`normalizePath()` doesn't work if the path contains Unicode characters that
cannot be represented in the current system locale, even if the file exists.
This function uses `normalize_directory()` to try to accomodate
symlink filenames. If it fails, it calls `fs::path_abs()` which
appears to be a more robust equivalent to `normalize_directory()`.
}
