% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{attach_output_format_dependency}
\alias{attach_output_format_dependency}
\title{Attach output format dependency to knit metadata}
\usage{
attach_output_format_dependency(dep, force = FALSE)
}
\arguments{
\item{dep}{A dependency created by \code{output_format_dependency()}}

\item{force}{attach even if \code{knitr::knit()} is not in progress
(default: \code{FALSE})}
}
\value{
invisible \code{NULL}
}
\description{
The attached dependency is merged to output format after calling the 
post_knit handler and before calling the pre_processor handler.
}
\examples{
# a function that attaches lua filters
f <- function() {
  dep <- output_format_dependency(
    "lua_filter",
    pre_processor = function(...) {
      pandoc_lua_filter_args(c("example1.lua", "example2.lua"))
    }
  )
  attach_output_format_dependency(dep)
}

}
