% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{available_templates}
\alias{available_templates}
\title{List available R Markdown template in a package}
\usage{
available_templates(package = "rmarkdown", full_path = FALSE)
}
\arguments{
\item{package}{Package to list template from. Default to \strong{rmarkdown}}

\item{full_path}{Set to \code{TRUE} to get the full path to the available templates}
}
\value{
A character vector of templates name to use within \code{\link[=draft]{draft()}}. If
\code{full_path = TRUE}, it returns the full path to the templates.
}
\description{
List the available template by name that can be used with \code{\link[=draft]{draft()}} to create
a new document for R Markdown from a package.
}
\examples{
# List rmarkdown templates & create a draft
available_templates()

# List rticles templates
available_templates("rticles")
}
