% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggress.R
\name{gginc}
\alias{gginc}
\title{Incremental graphs}
\usage{
gginc(loopVec, expr, print.expr = TRUE)
}
\arguments{
\item{loopVec}{vector of names or numbers for the incremental stages of the
graph. Usually, this will just be 1:n, where n is the number of stages.}

\item{expr}{the graphing command. Usually, but not necessarily, a (composite)
   ggplot command.}
}
\description{
Use this function to "wrap" a ggplot graph which will appear in incremental
stages in your presentation.
}
\details{
Usually, you want to begin the presentation with:
output:
  ioslides_presentation:
    incremental: true
    self_contained: false

... and the block with:
\`\`\`\{r, results='asis', error=FALSE, message=FALSE, warning=FALSE, echo=FALSE\}
}
\examples{
\donttest{
gginc(1:2, ggplot(mtcars,aes(y=mpg,x=wt)) + geom_point() + stages(geom_blank(),s2=geom_smooth()))
}
}

