\name{highlighter}
\alias{highlighter}
\title{Available syntax highlighers.}
\usage{
highlighter()
}
\value{
Character vector with available syntax highlighers
}
\description{
Get the available syntax highlighters (supported
highlighers are "default", "pygments", "kate",
"monochrome", "espresso", "zenburn", "haddock", and
"tango").
}
\details{
This is a convenience method for use in the signatures of
options functions (e.g. \code{\link{htmlOptions}} or
\code{\link{pdfOptions}}).
}

