% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_jquery}
\alias{html_dependency_jqueryui}
\alias{html_dependency_bootstrap}
\alias{html_dependency_tocify}
\alias{html_dependency_font_awesome}
\alias{html_dependency_ionicons}
\alias{html_dependency_pagedtable}
\alias{html_dependency_highlightjs}
\title{Provide common HTML dependencies for R Markdown formats}
\usage{
html_dependency_jquery()

html_dependency_jqueryui()

html_dependency_bootstrap(theme)

html_dependency_tocify()

html_dependency_font_awesome()

html_dependency_ionicons()

html_dependency_pagedtable()

html_dependency_highlightjs(highlight)
}
\arguments{
\item{theme}{Visual theme ("default", "cerulean", "cosmo", "cyborg", "darkly",
"flatly", "journal", "lumen", "readable", "sandstone", "simplex", "slate",
"spacelab", "superhero", "united", or "yeti"). Pass \code{NULL} for no
theme (in this case you can use the \code{css} parameter to add your own styles).
If Bootstrap 4+ is used (see \code{bootstrap_version} argument),
a handful of other themes are available ("lux", "minty", "pulse",
"sketchy", and "solar"). Moreover, with Bootstrap 4+, custom
\href{https://getbootstrap.com/docs/4.0/getting-started/theming/}{Bootstrap
themes} are supported via the \code{bs_theme_add_variables()} and
\code{bs_theme_add()} functions from the bootstraplib package
(use these functions from within the document).}

\item{highlight}{Highlighter to use}
}
\description{
These functions provide common HTML dependencies (e.g. jquery, bootstrap)
for re-use by other R Markdown formats.
}
