% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_notebook.R
\name{html_notebook}
\alias{html_notebook}
\title{Convert to an HTML notebook}
\usage{
html_notebook(code_folding = "show", highlight = "textmate", ...)
}
\arguments{
\item{code_folding}{Enable document readers to toggle the display of R code
chunks. Defaults to \code{"none"} which displays all code chunks (assuming
they were knit with \code{echo = TRUE}). Specify \code{"hide"} to hide all R
code chunks by default (users can show hidden code chunks either
individually or document-wide). Specify \code{"show"} to show all R code
chunks by default.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
"haddock", and "textmate". Pass \code{NULL} to prevent syntax highlighting.}

\item{...}{Optional arguments to be passed to \code{\link{html_document}}.}
}
\description{
Format for converting from R Markdown to an HTML notebook.
}

