% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_notebook_output.R
\name{html_notebook_metadata}
\alias{html_notebook_metadata}
\title{Generate R Notebook Metadata}
\usage{
html_notebook_metadata(iframe = TRUE)
}
\arguments{
\item{iframe}{Boolean; should output be shown in an \code{<iframe>}?}
}
\description{
A structured helper for the construction of metadata used by the
R Notebook output functions. See \code{\link{html_notebook_output}} for
more details.
}

