% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_notebook_output.R
\name{html_notebook_output}
\alias{html_notebook_output}
\alias{html_notebook_output_html}
\alias{html_notebook_output_png}
\title{Generate R Notebook Output}
\usage{
html_notebook_output_html(html, iframe = TRUE)

html_notebook_output_png(path = NULL, bytes = NULL, iframe = TRUE)
}
\arguments{
\item{html}{Arbitrary HTML content to insert.}

\item{iframe}{Boolean; should the generated HTML be shown in an iframe?}

\item{path}{A path to a file. For functions accepting both \code{path}
and \code{bytes}, if \code{bytes} is \code{NULL}, the bytewise contents
will be obtained by reading the file.}

\item{bytes}{The bytewise representation of content.}
}
\description{
Utilities for generating output for the \code{html_notebook} format,
through the \code{output_source} function attached to a
\code{\link{output_format}}.
}

