% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{knitr_options}
\alias{knitr_options}
\title{Knitr options for an output format}
\usage{
knitr_options(opts_knit = NULL, opts_chunk = NULL, knit_hooks = NULL,
  opts_hooks = NULL, opts_template = NULL)
}
\arguments{
\item{opts_knit}{List of package level knitr options (see
\code{\link[knitr:opts_knit]{opts_knit}})}

\item{opts_chunk}{List of chunk level knitr options (see
\code{\link[knitr:opts_chunk]{opts_chunk}})}

\item{knit_hooks}{List of hooks for R code chunks, inline R code, and output
(see \code{\link[knitr:knit_hooks]{knit_hooks}})}

\item{opts_hooks}{List of hooks for code chunk options
(see \code{\link[knitr:opts_hooks]{opts_hooks}})}

\item{opts_template}{List of templates for chunk level knitr options (see
\code{\link[knitr:opts_template]{opts_template}})}
}
\value{
An list that can be passed as the \code{knitr} argument
  of the \code{\link{output_format}} function.
}
\description{
Define the knitr options for an R Markdown output format.
}
\seealso{
\link{output_format}
}
