% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{knitr_options_pdf}
\alias{knitr_options_pdf}
\title{Knitr options for a PDF output format}
\usage{
knitr_options_pdf(fig_width, fig_height, fig_crop, dev = "pdf")
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{dev}{Graphics device to use for figure output (defaults to png)}
}
\value{
An list that can be passed as the \code{knitr} argument of the
  \code{\link{output_format}} function.
}
\description{
Define knitr options for an R Markdown output format that creates PDF output.
}
\seealso{
\link{knitr_options}, \link{output_format}
}

