% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_dependencies.R
\name{latex-dependencies}
\alias{latex-dependencies}
\alias{latex_dependency_tikz}
\title{Provide common LaTeX dependencies}
\usage{
latex_dependency_tikz(libraries, options = NULL, extra_lines = NULL)
}
\arguments{
\item{libraries}{A character vector of tikz libraries to load}

\item{options}{The LaTeX options for the package}

\item{extra_lines}{LaTeX code related to the package added to the preamble}
}
\description{
These functions provide common LaTeX dependencies (e.g. tikz)
for R Markdown formats that use LaTeX.
}
