\name{mathjaxURL}
\alias{mathjaxURL}
\title{MathJax URL}
\usage{
  mathjaxURL(version = "latest",
    config = "TeX-AMS-MML_HTMLorMML", https = FALSE)
}
\arguments{
  \item{version}{Version to use}

  \item{config}{Configuration to use}

  \item{https}{Use secure connection}
}
\value{
  URL to MathJax library
}
\description{
  Get the URL to the MathJax library.
}

