% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document.R
\name{navbar_html}
\alias{navbar_html}
\alias{navbar_links_html}
\title{Create a navbar HTML file from a navbar definition}
\usage{
navbar_html(navbar, bootstrap_version = c("3", "4", "4-3"))

navbar_links_html(links, bootstrap_version = c("3", "4", "4-3"))
}
\arguments{
\item{navbar}{Navbar definition}

\item{bootstrap_version}{major version of Bootstrap. Currently only 3 or 4 are supported.}

\item{links}{List of navbar links}
}
\value{
Path to temporary file with navbar definition
}
\description{
Create a navbar HTML file from a navbar definition
}
\keyword{internal}
