% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_paged.R
\name{paged_table}
\alias{paged_table}
\title{Create a table in HTML with support for paging rows and columns}
\usage{
paged_table(x, options = NULL)
}
\arguments{
\item{x}{a data frame to be rendered as a paged table.}

\item{options}{options for printing the paged table}
}
\description{
Create a table in HTML with support for paging rows and columns
}
