\name{pandoc-package}
\alias{pandoc-package}
\docType{package}
\title{Document conversion with pandoc}
\description{
Functions for converting documents with pandoc, including higher level functions
for rendering R Markdown documents in a variety of formats.
}

\author{
RStudio

Maintainer: RStudio <info@rstudio.com>
}

\keyword{ package }

