% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_citeproc_convert}
\alias{pandoc_citeproc_convert}
\title{Convert a bibliograpy file}
\usage{
pandoc_citeproc_convert(file, type = c("list", "json", "yaml"))
}
\arguments{
\item{file}{Bibliography file}

\item{type}{Conversion type}
}
\value{
For `type = "list"`, and R list. For `type = "json"` or `type = "yaml"`,
  a character vector with the specified format.
}
\description{
Convert a bibliography file (e.g. a BibTeX file) to an R list, JSON text,
  or YAML text
}
