% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc_copy.R
\name{pandoc_copy_data}
\alias{pandoc_copy_data}
\title{Copy a Pandoc's data file}
\usage{
pandoc_copy_data(data, output = data, overwrite = FALSE)
}
\arguments{
\item{data}{The name of Pandoc's data file (e.g., "reference.docx")}

\item{output}{The name of the output file. If \code{NULL} (default), the name
will be same as the \code{data} argument.}

\item{overwrite}{Whether or not to overwrite the existing file. Default to
\code{FALSE}.}
}
\value{
The name of the output file.
}
\description{
Copy a Pandoc's data file
}
\examples{
\dontrun{
pandoc_copy_data("reference.docx")
}

}
\seealso{
\url{https://pandoc.org/MANUAL.html#option--print-default-data-file}
}
