% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc_copy.R
\name{pandoc_copy_template}
\alias{pandoc_copy_template}
\title{Copy a template file for Pandoc}
\usage{
pandoc_copy_template(format, output = NULL, ...)
}
\arguments{
\item{format}{A format of the template file as a character (e.g.,
\code{html}) or as a result of a formatting function (e.g., \code{html_document()}).}

\item{output}{The name of the output file. If using \code{NULL} then the
output filename will be based on the \code{format} argument.}

\item{...}{Arguments passed to \code{file.copy()} when \format is a result of
a formatting function. Otherwise, they are ignored.}
}
\value{
The name of the output file.
}
\description{
Copy a template file for Pandoc
}
