% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_exec}
\alias{pandoc_exec}
\title{Get the path of the pandoc executable}
\usage{
pandoc_exec()
}
\description{
Returns the path of the pandoc executable used by functions in the the
\pkg{rmarkdown} package. This is the most recent version of pandoc found in
either the system path or shipped with RStudio.
}
\details{
See the
\href{https://pandoc.org/MANUAL.html}{pandoc manual}
for pandoc commands.
}
