% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_run}
\alias{pandoc_run}
\title{Run pandoc using the specified arguments.}
\usage{
pandoc_run(args, ...)
}
\arguments{
\item{args}{Command line arguments.}

\item{...}{Additional arguments to pass on to `system2()`}
}
\value{
Value from call to `system2()`
}
\description{
Low-level function to run pandoc using `system2()`.
}
