% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_self_contained_html}
\alias{pandoc_self_contained_html}
\title{Create a self-contained HTML document using pandoc.}
\usage{
pandoc_self_contained_html(input, output)
}
\arguments{
\item{input}{Input html file to create self-contained version of.}

\item{output}{Path to save output.}
}
\value{
(Invisibly) The path of the generated file.
}
\description{
Create a self-contained HTML document by base64 encoding images,
scripts, and stylesheets referred by the input document.
}
