\name{pandocConvert}
\alias{pandocConvert}
\title{Convert a document using pandoc}
\usage{
  pandocConvert(input, output = NULL, format = NULL,
    options = NULL, quiet = FALSE)
}
\arguments{
  \item{input}{Input file}

  \item{output}{Output file (if not specified then a
  default based on the specified format is chosen)}

  \item{format}{Pandoc format to convert to (defaults to
  HTML if not specified)}

  \item{options}{Command line options to pass to pandoc}

  \item{quiet}{\code{TRUE} to supress printing of the
  pandoc command line}
}
\value{
  The output file generated by the converstion
}
\description{
  Convert the \code{input} document to the specified
  \code{format} and write it to \code{output}.
}

