\name{pandocMarkdownOptions}
\alias{pandocMarkdownOptions}
\title{Pandoc options for R Markdown}
\usage{
  pandocMarkdownOptions(featureFlags = NULL)
}
\arguments{
  \item{featureFlags}{Additional flags for enabling and
  disabling markdown features supported during a
  conversion.}
}
\value{
  Character vector of pandoc options
}
\description{
  Get the pandoc command line options required to render R
  Markdown.
}

