\name{pandocPDFOptions}
\alias{pandocPDFOptions}
\title{Pandoc options for PDF rendering}
\usage{
  pandocPDFOptions(geometry = NULL)
}
\arguments{
  \item{geometry}{List of \code{LaTeX} geometry settings
  (optional)}
}
\value{
  Character vector of pandoc options
}
\description{
  Get pandoc command-line options required for converting R
  Markdown PDF.
}

