% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_notebook.R
\name{parse_html_notebook}
\alias{parse_html_notebook}
\title{Parse an HTML Notebook}
\usage{
parse_html_notebook(path, encoding = "UTF-8")
}
\arguments{
\item{path}{The path to an R Notebook file (with extension \code{.nb.html}).}

\item{encoding}{The document's encoding (assumend \code{"UTF-8"} by default).}
}
\description{
Parse an HTML notebook, retrieving annotation information
related to generated outputs in the document, as well as the
original R Markdown source document.
}
\details{
For more details on the HTML file format produced by
 \code{html_notebook}, see \href{http://rmarkdown.rstudio.com/r_notebook_format.html}{http://rmarkdown.rstudio.com/r_notebook_format.html}.
}

