% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pkg_file_lua}
\alias{pkg_file_lua}
\title{Get Lua filters full paths in a package}
\usage{
pkg_file_lua(filters = NULL, package = NULL)
}
\arguments{
\item{filters}{Character vector of filename for the Lua filter to retrieve
in \code{rmarkdown/lua} folder of the package. If \code{NULL}, all the
included lua filter will be returned. The extension can be ommited.}

\item{package}{Package name in which to look for the filters.}
}
\value{
Character vector of absolute file paths for the Lua filter from the
  package. If a filter is not found, it does not error but returns an empty
  path. The path will already be escaped correctly using
  \code{\link{pandoc_path_arg}} to be used by Pandoc.
}
\description{
Lua filters stored in a source package in \code{inst/rmarkdown/lua} will installed in
\code{rmarkdown/lua} in the package path. This function finds the full path of the
Lua filters on the system for installed packages.
}
\examples{
# get all filters
pkg_file_lua(package = "rmarkdown")
# get a specific filter
pkg_file_lua("pagebreak", package = "rmarkdown")
pkg_file_lua(c("pagebreak", "latex_divs"), package = "rmarkdown")
pkg_file_lua("pagebreak.lua", package = "rmarkdown")
}
