% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_site.R
\name{publish_site}
\alias{publish_site}
\title{Publish an R Markdown Website}
\usage{
publish_site(
  site_dir = ".",
  site_name = NULL,
  method = c("rsconnect"),
  server = NULL,
  account = NULL,
  render = TRUE,
  launch_browser = interactive()
)
}
\arguments{
\item{site_dir}{Directory containing website. Defaults to current working directory.}

\item{site_name}{Name for the site (names must be unique within an account). Defaults
to the `name` provided by the site generator (or to the name of the site_dir if
there is no `name` specified).}

\item{method}{Publishing method (currently only "rsconnect" is available)}

\item{render}{`TRUE` to render the site locally before publishing.}

\item{launch_browser}{If `TRUE`, the system's default web browser will be launched
automatically after the site is deployed. Defaults to `TRUE` in interactive sessions
only.}
}
\description{
Publish a website to RStudio Connect
}
\examples{
\dontrun{
library(rmarkdown)
publish_site()
}

}
