% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{html_dependency_jquery}
\alias{html_dependency_bootstrap}
\alias{html_dependency_jqueryui}
\alias{html_dependency_tocify}
\alias{html_dependency_font_awesome}
\alias{html_dependency_ionicons}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmldeps}{\code{\link[htmldeps]{html_dependency_jquery}}, \code{\link[htmldeps]{html_dependency_bootstrap}}, \code{\link[htmldeps]{html_dependency_jqueryui}}, \code{\link[htmldeps]{html_dependency_tocify}}, \code{\link[htmldeps]{html_dependency_font_awesome}}, \code{\link[htmldeps]{html_dependency_ionicons}}}
}}

