% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relative_to.R
\name{relative_to}
\alias{relative_to}
\title{Relative path utility function
Given a directory and a file, return a relative path from the directory to
the file, or the unmodified file path if the file does not appear to be in
the directory.}
\usage{
relative_to(dir, file)
}
\arguments{
\item{dir}{Directory}

\item{file}{File}
}
\value{
Relative path from the directory to the file (or the unmodified file
  path if the file does not appear to be in the directory).
}
\description{
Relative path utility function
Given a directory and a file, return a relative path from the directory to
the file, or the unmodified file path if the file does not appear to be in
the directory.
}

