% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{render_site}
\alias{default_site}
\alias{render_site}
\alias{site_config}
\alias{site_generator}
\title{Render multiple documents as a website}
\usage{
render_site(input = ".", output_format = "all", envir = parent.frame(),
  quiet = FALSE, encoding = getOption("encoding"), ...)

site_generator(input = ".", output_format = NULL,
  encoding = getOption("encoding"))

site_config(input, encoding = getOption("encoding"))

default_site(input, encoding = getOption("encoding"), ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory)}

\item{output_format}{R Markdown format to convert to.}

\item{envir}{The environment in which the code chunks are to be evaluated
during knitting (can use \code{\link{new.env}} to guarantee an empty new
environment).}

\item{quiet}{\code{TRUE} to supress messages and other output.}

\item{encoding}{The encoding of the input file; see \code{\link{file}}.}

\item{...}{Currently unused}
}
\value{
The name of the site output file (relative to the input directory).
}
\description{
Render all of the R Markdown documents within a directory as a website.
}

