\name{rmarkdownFormat}
\alias{rmarkdownFormat}
\title{Pandoc options for R Markdown}
\usage{
  rmarkdownFormat(options = NULL)
}
\arguments{
  \item{options}{Additional flags for enabling and
  disabling markdown features supported during a
  conversion.}
}
\value{
  Pandoc options
}
\description{
  Get the pandoc command line options required to render R
  Markdown.
}

