\name{rmarkdownOptions}
\alias{rmarkdownOptions}
\title{Pandoc options for R Markdown}
\usage{
  rmarkdownOptions(featureFlags = NULL)
}
\arguments{
  \item{featureFlags}{Additional flags for enabling and
  disabling markdown features supported during a
  conversion.}
}
\value{
  Pandoc options
}
\description{
  Get the pandoc command line options required to render R
  Markdown.
}

