% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{shiny_module}
\alias{shiny_module}
\title{Call a Shiny Module}
\usage{
shiny_module(module, ...)
}
\arguments{
\item{module}{A Shiny module server function}

\item{...}{Parameters to pass the module UI and server functions.}
}
\description{
This \code{shiny_module} function provides a shorthand syntax for calling
both the UI and server functions of a
\href{http://shiny.rstudio.com/articles/modules.html}{Shiny Module} within
a \code{runtime: shiny_prerendered} interactive document. Parameters passed
to the function will be forwarded as appropriate to the respective UI
and Server functions of the Shiny Module.
}
\details{
The \code{module} parmaeter is the name of the module's server
 function. The name of the module's UI function should use the server
 function name as it's base and then append either a `\code{UI}` or
 \code{_ui} suffix. For example, if the module server function is named
 \code{worldPhones} the UI function could be named \code{worldPhonesUI}.
}

