% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_prerendered.R
\name{shiny_prerendered_server_start_code}
\alias{shiny_prerendered_server_start_code}
\title{Get the server startup code for a shiny_prerendered server instance}
\usage{
shiny_prerendered_server_start_code(server_envir)
}
\arguments{
\item{server_envir}{Shiny server environment to get code for}
}
\description{
Get the server startup code for a shiny_prerendered server instance
}
\keyword{internal}

