% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_resources.R
\name{source_download_link}
\alias{source_download_link}
\title{Insert a link to download source code in an R Markdown file}
\usage{
source_download_link(text = "Download Source")
}
\arguments{
\item{text}{text to display in the link}
}
\description{
When using the \code{code_download} option in \code{\link{html_document}}, you
may use think funcion to insert a link in the document text to download
the source code or data.  This is useful when not using a theme, which
automatically places this link in the page menu
}
\examples{
\dontrun{
  To download the source code and data for this
  document, `r source_download_link("click here")`.
}
}

