% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggress.R
\name{stages}
\alias{stages}
\title{Variable portion within incremental graphs}
\usage{
stages(...)
}
\arguments{
\item{[positional_parameters]}{correct output for each stage, in sequential order.}

\item{[named_parameters_beginning_with_s]}{For instance, s5=0 says that beginning with
stage 5 (and until otherwise specified), the output should be 0.}

\item{[named_parameters_beginning_with_o]}{For instance, o5=0 says that for stage 5
   only, the output should be 0.}
}
\description{
Use this function for the portions which change inside a gginc expression.
}
\examples{
\donttest{
#4-stage plot: first points only, then add loess, then try linear fit, then return to loess.
gginc(1:4, ggplot(mtcars,aes(y=mpg,x=wt)) + geom_point() + stages(geom_blank(),s2=geom_smooth(),o3=geom_smo))

}
}

