% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tuftish_ebook.R
\name{tuftish_ebook}
\alias{tuftish_ebook}
\title{Tuftish e-Book format (PDF)}
\usage{
tuftish_ebook(toc = TRUE, toc_depth = 3, number_sections = TRUE,
  fig_width = 4, fig_height = 2.5, fig_crop = TRUE,
  highlight = "default", keep_tex = FALSE, latex_engine = "xelatex",
  includes = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{number_sections}{\code{TRUE} to number section headings}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
  "pdflatex", "lualatex", and "xelatex".  Note that lualatex may have
  problems with text spacing, and that pdflatex may have memory issues
  when using tikzDevice.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Template for creating an e-book based on the style of
Edward R. Tufte and Richard Feynman.
}

