% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_resources.R
\name{find_external_resources}
\alias{find_external_resources}
\title{Find External Resource References}
\usage{
find_external_resources(rmd_file, encoding = getOption("encoding"))
}
\arguments{
\item{rmd_file}{path to the R Markdown document to process}

\item{encoding}{the encoding of the R Markdown document}
}
\value{
A character vector containing the paths of the additional files. The
  paths appear as expressed in the document and are relative to the directory
  in which the document resides.
}
\description{
Given an R Markdown document, attempt to determine the set of additional
files needed in order to render and display the document.
}
\details{
This routine applies heuristics in order to scan a document for
  possible resource references.

  It looks for references to files expressed in Markdown (e.g.
  \code{![alt](img.png)}), in the document's YAML header, in raw HTML chunks,
  and as quoted strings in R code chunks (e.g. \code{read.csv("data.csv")}).

  Only resources that exist on disk and are contained in the document's
  directory (or a child thereof) are returned.
}

