% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_document.R
\name{github_document}
\alias{github_document}
\title{Convert to GitHub Flavored Markdown}
\usage{
github_document(fig_width = 7, fig_height = 5)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to GitHub Flavored Markdown.
}
\details{
The markdown syntax supported in this document is GitHub Flavored
  Markdown (more details are here:
  \href{https://help.github.com/articles/github-flavored-markdown/}{https://help.github.com/articles/github-flavored-markdown/}).
}
\note{
An HTML file is also generated for the purpose of locally previewing what
  the document will look like on GitHub.
}

