% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\docType{import}
\name{html-dependencies}
\alias{html-dependencies}
\alias{reexports}
\alias{html_dependency_jquery}
\alias{reexports}
\alias{html_dependency_bootstrap}
\alias{reexports}
\alias{html_dependency_jqueryui}
\alias{reexports}
\alias{html_dependency_tocify}
\alias{reexports}
\alias{html_dependency_font_awesome}
\alias{reexports}
\alias{html_dependency_ionicons}
\alias{html_dependency_pagedtable}
\alias{html_dependency_highlightjs}
\title{Provide common HTML dependencies for R Markdown formats}
\usage{
html_dependency_pagedtable()

html_dependency_highlightjs(highlight)
}
\arguments{
\item{highlight}{Highlighter to use}
}
\description{
These functions provide common HTML dependencies (e.g. jquery, bootstrap)
for re-use by other R Markdown formats.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmldeps}{\code{\link[htmldeps]{html_dependency_jquery}}, \code{\link[htmldeps]{html_dependency_bootstrap}}, \code{\link[htmldeps]{html_dependency_jqueryui}}, \code{\link[htmldeps]{html_dependency_tocify}}, \code{\link[htmldeps]{html_dependency_font_awesome}}, \code{\link[htmldeps]{html_dependency_ionicons}}}
}}

