% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_pagedtable}
\alias{html_dependency_highlightjs}
\title{Provide common HTML dependencies for R Markdown formats}
\usage{
html_dependency_pagedtable()

html_dependency_highlightjs(highlight)
}
\arguments{
\item{highlight}{Highlighter to use}
}
\description{
These functions provide common HTML dependencies (e.g. jquery, bootstrap)
for re-use by other R Markdown formats.
}
