% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_bootstrap}
\alias{html_dependency_font_awesome}
\alias{html_dependency_ionicons}
\alias{html_dependency_jquery}
\alias{html_dependency_jqueryui}
\alias{html_dependency_tocify}
\title{Provide common HTML dependencies for R Markdown formats}
\usage{
html_dependency_jquery()

html_dependency_bootstrap(theme)

html_dependency_jqueryui()

html_dependency_tocify()

html_dependency_font_awesome()

html_dependency_ionicons()
}
\arguments{
\item{theme}{Visual theme ("default", "cerulean", "journal", "flatly",
"readable", "spacelab", "united", "cosmo", "lumen", "paper", "sandstone",
"simplex", or "yeti"). Pass \code{NULL} for no theme (in this case you can
use the \code{css} parameter to add your own styles).}
}
\description{
These functions provide common HTML dependencies (e.g. jquery, bootstrap)
for re-use by other R Markdown formats.
}

