% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_jquery}
\alias{html_dependency_jqueryui}
\alias{html_dependency_bootstrap}
\alias{html_dependency_tocify}
\alias{html_dependency_font_awesome}
\alias{html_dependency_ionicons}
\alias{html_dependency_pagedtable}
\alias{html_dependency_highlightjs}
\title{Provide common HTML dependencies for R Markdown formats}
\usage{
html_dependency_jquery()

html_dependency_jqueryui()

html_dependency_bootstrap(theme)

html_dependency_tocify()

html_dependency_font_awesome()

html_dependency_ionicons()

html_dependency_pagedtable()

html_dependency_highlightjs(highlight)
}
\arguments{
\item{theme}{Visual theme ("default", "cerulean", "journal", "flatly",
"darkly", "readable", "spacelab", "united", "cosmo", "lumen", "paper",
"sandstone", "simplex", or "yeti"). Pass \code{NULL} for no theme (in this
case you can use the \code{css} parameter to add your own styles).}

\item{highlight}{Highlighter to use}
}
\description{
These functions provide common HTML dependencies (e.g. jquery, bootstrap)
for re-use by other R Markdown formats.
}
