% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_jquery}
\alias{html_dependency_jqueryui}
\alias{html_dependency_bootstrap}
\alias{html_dependency_tocify}
\alias{html_dependency_font_awesome}
\alias{html_dependency_ionicons}
\alias{html_dependency_pagedtable}
\alias{html_dependency_highlightjs}
\title{Provide common HTML dependencies for R Markdown formats}
\usage{
html_dependency_jquery()

html_dependency_jqueryui()

html_dependency_bootstrap(theme)

html_dependency_tocify()

html_dependency_font_awesome()

html_dependency_ionicons()

html_dependency_pagedtable()

html_dependency_highlightjs(highlight)
}
\arguments{
\item{theme}{One of the following:
\itemize{
\item A \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object (or a list of \code{\link[bslib:bs_theme]{bslib::bs_theme()}} argument values)
\itemize{
\item Use this option for custom themes using Bootstrap 4 or 3.
\item In this case, any \code{.scss}/\code{.sass} files provided to the \code{css}
parameter may utilize the \code{theme}'s underlying Sass utilities
(e.g., variables, mixins, etc).
}
\item \code{NULL} for no theme (i.e., no \code{\link[=html_dependency_bootstrap]{html_dependency_bootstrap()}}).
\item A character string specifying a \href{https://bootswatch.com/3/}{Bootswatch 3}
theme name (for backwards-compatibility).
}}

\item{highlight}{Highlighter to use}
}
\description{
These functions provide common HTML dependencies (e.g. jquery, bootstrap)
for re-use by other R Markdown formats.
}
