% Generated by roxygen2 (4.0.2): do not edit by hand
\name{html_vignette}
\alias{html_vignette}
\title{Convert to an HTML vignette.}
\usage{
html_vignette(fig_width = 3, fig_height = 3, css = NULL, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{css}{One or more css files to include}

\item{...}{Additional function arguments to pass to the base R Markdown HTML
  output formatter}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
A HTML vignette is a lightweight alternative to \code{\link{html_document}}
suitable for inclusion in packages to be released to CRAN. It reduces the
size of a basic vignette from 100k to around 10k.
}
\details{
Compared to \code{html_document}, it:

\itemize{
  \item never uses retina figures
  \item has a smaller default figure size
  \item uses a custom css stylesheet
 }
}

