% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{knit_params_ask}
\alias{knit_params_ask}
\title{Run a shiny application asking for parameter configuration for the given document.}
\usage{
knit_params_ask(file = "index.Rmd", params = NULL, shiny_args = NULL)
}
\arguments{
\item{file}{Path to the R Markdown document with configurable parameters.}

\item{params}{A named list of optional parameter overrides used in place of the document defaults.}

\item{shiny_args}{Additional arguments to \code{\link[shiny:runApp]{runApp}}.}
}
\value{
named list with overridden parameter names and value.
}
\description{
Run a shiny application asking for parameter configuration for the given document.
}

