% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{knit_params_ask}
\alias{knit_params_ask}
\title{Run a shiny application asking for parameter configuration for the given document.}
\usage{
knit_params_ask(file = NULL, input_lines = NULL, params = NULL,
  shiny_args = NULL, save_caption = "Save",
  encoding = getOption("encoding"))
}
\arguments{
\item{file}{Path to the R Markdown document with configurable parameters.}

\item{input_lines}{Content of the R Markdown document. If \code{NULL}, the contents of \code{file} will be read.}

\item{params}{A named list of optional parameter overrides used in place of the document defaults.}

\item{shiny_args}{Additional arguments to \code{\link[shiny:runApp]{runApp}}.}

\item{save_caption}{Caption to use use for button that saves/confirms parameters.}

\item{encoding}{The encoding of the input file; see \code{\link{file}}.}
}
\value{
named list with overridden parameter names and value.
}
\description{
Run a shiny application asking for parameter configuration for the given document.
}

