% Generated by roxygen2 (4.0.2): do not edit by hand
\name{knitr_options_html}
\alias{knitr_options_html}
\title{Knitr options for an HTML output format}
\usage{
knitr_options_html(fig_width, fig_height, fig_retina, keep_md)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays (defaults to 2 when
\code{fig_caption} is \code{FALSE}, which currently works for all widely
used retina displays). Set to \code{NULL} to prevent retina scaling. Note
that this will always be \code{NULL} when \code{keep_md} is specified (this
is because \code{fig_retina} relies on outputting HTML directly into the
markdown document).}

\item{keep_md}{Keep the markdown file generated by knitting.}
}
\value{
An list that can be passed as the \code{knitr} argument of the
  \code{\link{output_format}} function.
}
\description{
Define knitr options for an R Markdown output format that creates
HTML output.
}
\seealso{
\link{knitr_options}, \link{output_format}
}

