% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{knitr_options_pdf}
\alias{knitr_options_pdf}
\title{Knitr options for a PDF output format}
\usage{
knitr_options_pdf(fig_width, fig_height, fig_crop, dev = "pdf")
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{fig_crop}{Whether to crop PDF figures with the command
\command{pdfcrop}. This requires the tools \command{pdfcrop} and
\command{ghostscript} to be installed. By default, \code{fig_crop = TRUE}
if these two tools are available.}

\item{dev}{Graphics device to use for figure output (defaults to png)}
}
\value{
An list that can be passed as the \code{knitr} argument of the
  \code{\link{output_format}} function.
}
\description{
Define knitr options for an R Markdown output format
that creates PDF output.
}
\seealso{
\link{knitr_options}, \link{output_format}
}
