% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{metadata}
\alias{metadata}
\title{The YAML metadata of the current R Markdown document}
\format{\preformatted{ list()
}}
\usage{
metadata
}
\description{
The object \code{metadata} stores the YAML metadata of the current R Markdown
document as a list, which you may use in the R code chunks, e.g.
\code{metadata$title} (the title of the document), \code{metadata$author},
and \code{metadata$foo} (if you have a YAML field named \code{foo}), etc.
}
\keyword{datasets}

