% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{The YAML metadata of the current R Markdown document}
\format{\preformatted{ list()
}}
\description{
The object \code{metadata} stores the YAML metadata of the current R Markdown
document as a list, which you may use in the R code chunks, e.g.
\code{rmarkdown::metadata$title} (the title of the document),
\code{rmarkdown::metadata$author}, and \code{rmarkdown::metadata$foo} (if you
have a YAML field named \code{foo}), etc.
}
\examples{
rmarkdown::metadata
}
\keyword{datasets}

