% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_supporting_files}
\alias{render_supporting_files}
\title{Render supporting files for an input document}
\usage{
render_supporting_files(from, files_dir, rename_to = NULL)
}
\arguments{
\item{from}{The directory from which the files should be copied.}

\item{files_dir}{The directory that will receive the copied files.}

\item{rename_to}{An option to rename the source directory after the copy
operation is complete.}
}
\value{
The relative path to the supporting files. This path is suitable
for inclusion in HTML\code{href} and \code{src} attributes.
}
\description{
Render (copy) required supporting files for an input document to the
\code{_files} directory that is associated with the document.
}
