% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_prerendered.R
\name{shiny_prerendered_clean}
\alias{shiny_prerendered_clean}
\title{Clean prerendered content for the specified Rmd input file}
\usage{
shiny_prerendered_clean(input)
}
\arguments{
\item{input}{Rmd input file to clean content for}
}
\description{
Remove the associated html file and supporting _files directory
for a shiny_prerendered documet.
}

