% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_site.R
\name{site_resources}
\alias{site_resources}
\title{Determine website resource files for a directory}
\usage{
site_resources(site_dir, include = NULL, exclude = NULL,
  recursive = FALSE)
}
\arguments{
\item{site_dir}{Site directory to analyze}

\item{include}{Additional files to include (glob wildcards supported)}

\item{exclude}{Files to exclude  (glob wildcards supported)}

\item{recursive}{`TRUE` to return a full recursive file listing; `FALSE` to
just provide top-level files and directories.}
}
\value{
Character vector of files and directories to copy
}
\description{
Determine which files within a given directory should be copied in
order to serve a website from the directory. Attempts to automatically
exclude source, data, hidden, and other files not required to serve
website content.
}
