% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{yaml_front_matter}
\alias{yaml_front_matter}
\title{Parse the YAML front matter from a file}
\usage{
yaml_front_matter(input)
}
\arguments{
\item{input}{Input file (Rmd or plain markdown)}
}
\description{
Parse the YAML front matter from a file
}
\keyword{internal}
